new feature: process sequences with N's
Previously, only sequences without N's were
allowed to be processed, e.g. to determine the
motif score and motif matches.
Now, it is possible to scan sequences with N's.
In case, N's are encountered, the scores at the
respective positions evaluate to NaN.
Therefore, it is not possible to determine
scoreSequence, motifHits, scoreProfile and motifHitProfile
using N-containing sequences.
N-containing sequences are still not permitted
for the enrichment test.